requirements.txt
playwright
python-dotenv
playwright-stealth==1.0.6
